refactor(act-rules): `ACTRules` is now `ACTRulesRunner`, `ACTRulesModule` is now `ACTRules`
This should better align meaning with readability of external API.
